// app/hooks/types.ts
export enum BuildType {
  STAGING = "STAGING",
  SANDBOX = "SANDBOX",
  PRODUCTION = "PRODUCTION",
}
// ...
export interface ApiResponse<T> {
  data: T;
  status: string;
}

export type Callback<T> = (result: T | null, error: Error | null) => void;

export interface AuthDetails {
  authToken: string;
  refreshToken: string;
  deviceToken: string;
}
// ...
export interface Portfolio {
  token_name: string;
  token_image: string;
  token_address: string;
  network_name: string;
  quantity: string;
  amount_in_inr: string;
}

export interface PortfolioData {
  tokens: Portfolio[];
  total: number;
}

export interface Token {
  token_name: string;
  token_address: string;
  network_name: string;
}

export interface TokensData {
  tokens: Token[];
}

export interface User {
  email: string;
  user_id: string;
  created_at: string;
  freezed: string;
  freeze_reason: string;
}

export interface Wallet {
  network_name: string;
  address: string;
  success: boolean;
}

export interface WalletData {
  wallets: Wallet[];
}

export interface RawTransactionStatusQuery {
  order_id: string;
}

export interface RawTransactionStatus {
  order_id: string;
  network_name: string;
  status: string;
  transaction_hash: string;
}

export interface RawTransactionStatusData {
  total: number;
  jobs: RawTransactionStatus[];
}

export interface TransferTokens {
  network_name: string;
  token_address: string;
  quantity: string;
  recipient_address: string;
}

export interface TransferTokensData {
  orderId: string;
}